Fix end-of-options parsing, make argument retrieval a bit cleaner
The pattern '--*' in the argument parser was matching the
end-of-options string '--'. As a result, the string was collapsed to a
blank string in the parser but not in the argument list itself. This
triggered the parser to stop argument parsing, but without removing the
'--' string from the argument list, resulting in it being sent as a
command to the Tor daemon.

To fix, change the patterns '--*' and '-*' in the parser to '--?*' and
'-?*', respectively. This makes the parser collapse '--' to '-', which
it later recognizes as an end-of-options indicator. This causes the
parser to remove '--' from the argument list before stopping argument
parsing.

Also, stop assuming '$2' is always defined. We don't use 'set -u' in
tor-ctrl today, but we might in the future.